Added an example implementation and a few fixes
ADDED: a directory example/ and implemented Google Authentication using tomcat/spring. More example with Facebook, Github are on it's way
FIXED: a problem in AuthenticationService was invalidating the session of the user who signed up for the first time.
FIXED: we would sometimes get invalid_grant error. So handled invalidating the session when we get that message